# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Rust Web3 Risk Platform API
  version: "0.1.0"
  description: >
    Minimal reference for the Axum API surface, including auth (SIWE),
    wallets, strategies/backtests, and portfolio endpoints. Authenticated
    endpoints accept either Authorization bearer tokens or the httpOnly
    cookies issued during login.
servers:
  - url: http://localhost:8081
paths:
...
  /api/portfolio/{wallet_id}:
    get:
      security:
        - bearerAuth: []
      summary: Retrieve the latest portfolio snapshot
      parameters:
        - in: path
          name: wallet_id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: Snapshot returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioSnapshot"
        "404":
          description: Wallet not found
  /api/portfolio/{wallet_id}/history:
    get:
      security:
        - bearerAuth: []
      summary: Retrieve portfolio history
      parameters:
        - in: path
          name: wallet_id
          required: true
          schema:
            format: uuid
            type: string
        - in: query
          name: limit
          description: Maximum entries to return (default 50, max 500)
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: History returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioSnapshot"
        "404":
          description: Wallet not found
  /api/portfolio/{wallet_id}/snapshots:
    get:
      security:
        - bearerAuth: []
      summary: Retrieve portfolio snapshots within days
      parameters:
        - in: path
          name: wallet_id
          required: true
          schema:
            format: uuid
            type: string
        - in: query
          name: days
          description: How many days to include (default 7, max 30)
          schema:
            type: integer
            minimum: 1
            maximum: 30
      responses:
        "200":
          description: Snapshots returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioSnapshot"
        "404":
          description: Wallet not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PortfolioSnapshot:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        positions:
          type: array
          items:
            $ref: "#/components/schemas/Position"
        total_usd_value:
          type: number
        timestamp:
          type: string
          format: date-time
      required:
        - wallet_id
        - positions
        - total_usd_value
        - timestamp
    Position:
      type: object
      properties:
        asset_symbol:
          type: string
        amount:
          type: number
        usd_value:
          type: number
      required:
        - asset_symbol
        - amount
        - usd_value